Sort fixtures that compile to localized acceptance/unit conftest respectively; Faster unit test suite.
